+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to this project will be documented in this file.
 - Add tooltips to tag buttons in the checks list screen (#911)
 - Improve Email - Keywoard Filtering docs (by @mmomjian)
 - Split the grace time input field into value/unit input group (#945, by @mickBoat00)
+- Add a system check to warn about MariaDB UUID migration (#929)
 
 ### Bug Fixes
 - Increase uWSGI buffer size to allow requests with large cookies (#925)